+++ HISTORY.md
@@ -1,9 +1,12 @@
 # Snap! (BYOB) History
 
 ## in development:
+* **Notable Fixes:**
+    * fixed 'distribution' list selector support for nested atomic lists
 
 2023-11-21
 * new dev version for v9.0.14
+* threads: fixed 'distribution' list selector support for nested atomic lists
 
 ## 9.0.13:
 * **Notable Fixes:**